+++ src/main/kotlin/dev/mizarc/waystonewarps/application/results/UpdateWarpNameResult.kt
@@ -1,0 +1,8 @@
+package dev.mizarc.waystonewarps.application.results
+
+enum class UpdateWarpNameResult {
+    SUCCESS,
+    WARP_NOT_FOUND,
+    NAME_ALREADY_TAKEN,
+    NAME_BLANK
+}