+++ src/main/kotlin/dev/mizarc/waystonewarps/domain/discoveries/Discovery.kt
@@ -9,11 +9,11 @@ import java.util.*
  *
  * @param warpId Unique identifier for the warp.
  * @param playerId Unique identifier for the player.
- * @param firstDiscoveredTime Time the waystone was discovered.
+ * @param discoveredTime Time the waystone was discovered.
  * @param lastVisitedTime Time the waystone was last teleported to.
  * @param isFavourite If the waystone is marked as favourite.
  */
-class Discovery(var warpId: UUID, var playerId: UUID, var firstDiscoveredTime: Instant,
+class Discovery(var warpId: UUID, var playerId: UUID, var discoveredTime: Instant,
                 var lastVisitedTime: Instant, var isFavourite: Boolean) {
 
     /**
@@ -21,8 +21,8 @@ class Discovery(var warpId: UUID, var playerId: UUID, var firstDiscoveredTime: I
      *
      * @param warpId Unique identifier for the warp.
      * @param playerId Unique identifier for the player.
-     * @param firstDiscoveredTime Time the waystone was discovered.
+     * @param discoveredTime Time the waystone was discovered.
      */
-    constructor(warpId: UUID, playerId: UUID, firstDiscoveredTime: Instant): this(
-        warpId, playerId, firstDiscoveredTime, firstDiscoveredTime, false)
+    constructor(warpId: UUID, playerId: UUID, discoveredTime: Instant): this(
+        warpId, playerId, discoveredTime, discoveredTime, false)
 }