+++ src/main/kotlin/dev/mizarc/waystonewarps/interaction/menus/Menu.kt
@@ -3,5 +3,8 @@ package dev.mizarc.waystonewarps.interaction.menus
 import org.bukkit.entity.Player
 
 interface Menu {
-    fun open(player: Player)
+    fun open()
+    fun passData(data: Any?) {
+        // Default implementation does nothing
+    }
 }