+++ src/main/kotlin/dev/mizarc/waystonewarps/interaction/menus/management/WarpRenamingMenu.kt
@@ -4,6 +4,7 @@ import com.github.stefvanschie.inventoryframework.gui.GuiItem
 import com.github.stefvanschie.inventoryframework.gui.type.AnvilGui
 import com.github.stefvanschie.inventoryframework.pane.StaticPane
 import dev.mizarc.waystonewarps.application.actions.management.UpdateWarpName
+import dev.mizarc.waystonewarps.application.results.UpdateWarpNameResult
 import dev.mizarc.waystonewarps.domain.warps.Warp
 import dev.mizarc.waystonewarps.interaction.menus.Menu
 import dev.mizarc.waystonewarps.interaction.menus.MenuNavigator
@@ -15,14 +16,24 @@ import org.bukkit.inventory.ItemStack
 import org.koin.core.component.KoinComponent
 import org.koin.core.component.inject
 
-class WarpRenamingMenu(private val menuNavigator: MenuNavigator, private val warp: Warp): Menu, KoinComponent {
+class WarpRenamingMenu(private val player: Player, private val menuNavigator: MenuNavigator,
+                       private val warp: Warp): Menu, KoinComponent {
     private val updateWarpName: UpdateWarpName by inject()
-    private var nameAttempt = ""
 
-    override fun open(player: Player) {
+    private var name = ""
+    private var isConfirming = false
+
+    override fun open() {
         // Create homes menu
         val gui = AnvilGui("Renaming Warp")
         gui.setOnTopClick { guiEvent -> guiEvent.isCancelled = true }
+        gui.setOnNameInputChanged { newName ->
+            if (!isConfirming) {
+                name = newName
+            } else {
+                isConfirming = false
+            }
+        }
 
         // Add lodestone menu item
         val firstPane = StaticPane(0, 0, 1, 1)
@@ -34,34 +45,48 @@ class WarpRenamingMenu(private val menuNavigator: MenuNavigator, private val war
         gui.firstItemComponent.addPane(firstPane)
 
         // Add message menu item if name is already taken
-        if (nameAttempt != "") {
-            val secondPane = StaticPane(0, 0, 1, 1)
-            val paperItem = ItemStack(Material.PAPER)
-                .name("That name has already been taken")
-            val guiPaperItem = GuiItem(paperItem) { guiEvent -> guiEvent.isCancelled = true }
-            secondPane.addItem(guiPaperItem, 0, 0)
-            gui.secondItemComponent.addPane(secondPane)
-        }
+        val secondPane = StaticPane(0, 0, 1, 1)
+        gui.secondItemComponent.addPane(secondPane)
 
         // Add confirm menu item.
         val thirdPane = StaticPane(0, 0, 1, 1)
         val confirmItem = ItemStack(Material.NETHER_STAR).name("Confirm")
         val confirmGuiItem = GuiItem(confirmItem) { guiEvent ->
             // Go back to edit menu if the name hasn't changed
-            if (gui.renameText == warp.name) {
-                menuNavigator.goBack(player)
+            if (name == warp.name) {
+                menuNavigator.goBack()
                 return@GuiItem
             }
 
-            // Stay on menu if the name is already taken
-            val result = updateWarpName.execute(warp.id, gui.renameText)
-            if (result.isFailure) {
-                nameAttempt = gui.renameText
-                open(player)
-                return@GuiItem
+            // Attempt renaming
+            val result = updateWarpName.execute(warp.id, player.uniqueId, name)
+            when (result) {
+                UpdateWarpNameResult.SUCCESS -> menuNavigator.goBack()
+                UpdateWarpNameResult.WARP_NOT_FOUND -> {
+                    val paperItem = ItemStack(Material.PAPER)
+                        .name("The warp you are trying to rename does not exist anymore")
+                    val guiPaperItem = GuiItem(paperItem)
+                    secondPane.addItem(guiPaperItem, 0, 0)
+                    lodestoneItem.name(name)
+                    isConfirming = true
+                    gui.update()
+                }
+                UpdateWarpNameResult.NAME_ALREADY_TAKEN -> {
+                    val paperItem = ItemStack(Material.PAPER)
+                        .name("That name has already been taken")
+                    val guiPaperItem = GuiItem(paperItem) {guiEvent ->
+                        secondPane.removeItem(0, 0)
+                        lodestoneItem.name(name)
+                        isConfirming = true
+                        gui.update()
+                    }
+                    secondPane.addItem(guiPaperItem, 0, 0)
+                    lodestoneItem.name(name)
+                    isConfirming = true
+                    gui.update()
+                }
+                UpdateWarpNameResult.NAME_BLANK -> menuNavigator.goBack()
             }
-
-           menuNavigator.goBack(player)
         }
 
         thirdPane.addItem(confirmGuiItem, 0, 0)